api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
pydantic>=2.10.0
anyio>=4.8.0
python-multipart==0.0.6
python-dotenv>=1.0.0
PyYAML==6.0.1
boto3>=1.34.0
requests>=2.31.0
tenacity>=8.3.0
orjson>=3.10.0
shortuuid>=1.0.12
typer>=0.12.3
google-auth>=2.31.0
opentelemetry-api>=1.36.0
opentelemetry-sdk>=1.36.0
opentelemetry-instrumentation-fastapi>=0.57b0
opentelemetry-exporter-prometheus>=0.57b0
prometheus-client>=0.22.1
